+++ src/features/weather/index.ts
@@ -1,1 +1,2 @@
 export * from "./CurrentWeather"
+export * from "./CurrentWeatherFallback"